feature/search_page2 (#5)
* pod RxReachability, Kingfisher 추가

* IndicatorView 추가

* Reachability 추가

* Network Disconnect AlertController 추가

* SearchTableView Cell UI 추가

* SearchTableViewCell이 AppItem만 들고있도록 수정

* setUp 추가, 썸네일 추가

* Reachability 삭제, NetworkReachabilityManager 사용

* 함수 위치 조정

+++ AppStoreClone/Model/AppItemResponse.swift
@@ -24,7 +24,7 @@ public struct AppItem: Decodable {
     public let genres: [String]
     public let price: Float
     public let description: String
-    public let userRatingCount: Int
+    public let userRatingCount: Double
 }
 
 public struct AppItemResponse: Decodable {
